update schema, update register, login multiple device, logout

+++ src/auth/auth.service.ts
@@ -10,11 +10,19 @@ import { InjectMapper } from '@automapper/nestjs'
 import { UserEntity } from '../users/users.entity'
 import { sentVerifyEmail } from './utils/email'
 import { VerifyTokenDto } from '../users/dto/verifyToken.dto'
+import { LoginUserDto } from 'src/users/dto/login.dto'
+import { TokenService } from '../token/token.service'
+import { TokenEntity } from 'src/token/token.entity'
+import { TokenLoginDto } from './dto/token.login.dto'
+import { randomUUID } from 'crypto'
+import { JtiDto } from '../token/dto/jti.dto'
+
 @Injectable()
 export class AuthService {
   constructor(
     private usersService: UsersService,
     private jwtService: JwtService,
+    private tokenService: TokenService,
     @InjectMapper() private readonly classMapper: Mapper
   ) {}
 
@@ -23,16 +31,19 @@ export class AuthService {
 
     let userInfo = this.classMapper.map(registration, RegisterUserDto, UserEntity)
     userInfo.createdAt = Date.now().toString()
-    userInfo.isActive = false
 
     try {
       const user = await this.usersService.create(userInfo)
-      user.verifyToken = this.jwtService.sign({ id: user.id, createdAt: user.createdAt })
+      user.verifyToken = this.jwtService.sign(
+        { id: user.id, createdAt: user.createdAt },
+        { secret: process.env.JWT_SECRET_KEY, expiresIn: process.env.JWT_EXPIRED_TIME }
+      )
       await this.usersService.updateById(user, { verifyToken: user.verifyToken })
       sentVerifyEmail(userInfo.email, user.verifyToken)
 
       return this.classMapper.map(user, UserEntity, ReadUserDto)
     } catch (error) {
+      console.log(error)
       if (error instanceof Prisma.PrismaClientKnownRequestError) {
         if (error.code == 'P2002') {
           throw new HttpException('Email exist', HttpStatus.BAD_REQUEST)
@@ -42,44 +53,43 @@ export class AuthService {
     }
   }
 
-  async verify(tokenQuery: VerifyTokenDto): Promise<ReadUserDto | any> {
+  async verify(tokenQuery: VerifyTokenDto): Promise<ReadUserDto> {
     const token = tokenQuery.token
     let userEntity = this.classMapper.map(tokenQuery, VerifyTokenDto, UserEntity)
+    let user = await this.usersService.findByToken(userEntity)
+
+    /* Not found user with token */
+    if (!user) {
+      throw new HttpException("Verify token doesn't exist", HttpStatus.UNAUTHORIZED)
+    }
+
     try {
-      /* Case 1: JWT verify token does not expire */
-      let payload = this.jwtService.verify(token)
-      let user = await this.usersService.findByToken(userEntity)
-
-      /* If account does not active */
-      if (!user.isActive) {
-        user = await this.usersService.updateById(user, {
-          verifiedAt: Date.now().toString(),
-          isActive: true,
-        })
-      }
-
-      return this.classMapper.map(user, UserEntity, ReadUserDto)
+      var payload = this.jwtService.verify(token, { secret: process.env.JWT_SECRET_KEY })
     } catch (error) {
-      /* Case 2: If JWT verify token expired */
-      let user = await this.usersService.findByToken(userEntity)
+      throw error
+    }
 
-      /* Can't find account with token */
-      if (!user) {
-        throw new HttpException('Token does not exist', HttpStatus.UNAUTHORIZED)
-      }
-
-      /* Exist account match token but this account didn't verify */
+    /* Case 1: Token expired */
+    if (payload.exp * 1000 < Date.now()) {
+      /* Delete user info if user don't active */
       if (!user.isActive) {
         await this.usersService.deleteById(user)
       }
-
-      /* Exist account match token and this account verified */
-      throw new HttpException('Token expired', HttpStatus.UNAUTHORIZED)
+      throw new HttpException('Verify token expired', HttpStatus.UNAUTHORIZED)
     }
+
+    /* Case 2: Token doesn't expire */
+    if (!user.isActive) {
+      user = await this.usersService.updateById(user, {
+        verifiedAt: Date.now().toString(),
+        isActive: true,
+      })
+    }
+    return this.classMapper.map(user, UserEntity, ReadUserDto)
   }
 
-  async validateUser(validateUserData: RegisterUserDto): Promise<ReadUserDto> {
-    const userEntity = this.classMapper.map(validateUserData, RegisterUserDto, UserEntity)
+  async validateUser(validateUserData: LoginUserDto): Promise<TokenLoginDto> {
+    const userEntity = this.classMapper.map(validateUserData, LoginUserDto, UserEntity)
     const user = await this.usersService.findByEmail(userEntity)
     if (!user) {
       throw new HttpException('Email not exist', HttpStatus.UNAUTHORIZED)
@@ -94,20 +104,48 @@ export class AuthService {
 
     let isMatch = await bcrypt.compare(validateUserData.password, user.hash)
     if (isMatch) {
-      user.lastLogin = Date.now().toString()
-      await this.usersService.updateById(user, { lastLogin: user.lastLogin })
-      return this.classMapper.map(user, UserEntity, ReadUserDto)
+      let jti = randomUUID()
+      const token = this.getJwtToken(user.id, jti)
+      const { iat, exp } = this.jwtService.verify(token.refreshToken, {
+        secret: process.env.REFRESH_TOKEN_SECRET_KEY,
+      })
+
+      let tokenEntity = new TokenEntity(
+        user.id,
+        jti,
+        (iat * 1000).toString(),
+        (exp * 1000).toString()
+      )
+
+      await this.tokenService.create(tokenEntity)
+      return token
     }
     throw new HttpException('Password not match', HttpStatus.UNAUTHORIZED)
   }
 
-  getJwtToken(id: string) {
-    const payload = { id }
-    const token = this.jwtService.sign(payload)
-    return `Authentication=${token}; HttpOnly; Path=/; Max-Age=${process.env.JWT_EXPIRED_TIME}`
+  getJwtToken(id: string, uuid: string) {
+    const payload = { userId: id }
+    const accessToken = this.jwtService.sign(payload, {
+      secret: process.env.ACCESS_TOKEN_SECRET_KEY,
+      expiresIn: process.env.ACCESS_TOKEN_EXPIRED_TIME,
+      jwtid: uuid,
+    })
+    const refreshToken = this.jwtService.sign(payload, {
+      secret: process.env.REFRESH_TOKEN_SECRET_KEY,
+      expiresIn: process.env.REFRESH_TOKEN_EXPIRED_TIME,
+      jwtid: uuid,
+    })
+    return { accessToken, refreshToken }
   }
 
-  getCookieForLogOut() {
-    return `Authentication=; HttpOnly; Path=/; Max-Age=0`
+  async logout(jtiDto: JtiDto): Promise<TokenEntity> {
+    let tokenEntity = this.classMapper.map(jtiDto, JtiDto, TokenEntity)
+    console.log(tokenEntity)
+    tokenEntity = await this.tokenService.findByJtiAndId(tokenEntity)
+    console.log(tokenEntity)
+    if (!tokenEntity) {
+      throw new HttpException('Unauthorized', HttpStatus.UNAUTHORIZED)
+    }
+    return await this.tokenService.deleteById(tokenEntity)
   }
 }
